// src/lib/schema.ts
export type Repo = {
  id: string
  owner: string
  lang: string
  nodes: number
  edges: number
  funcs: number
  methods: number
  types: number
  interfaces: number
  vars: number
  files: number
  color: string
}